+++ src/N98/Magento/Application/ConfigurationLoader.php
@@ -157,11 +157,11 @@ class ConfigurationLoader
     protected function loadDistConfig(array $initConfig)
     {
         if ($this->_distConfig == null) {
-            $this->_distConfig = Yaml::parse(__DIR__ . '/../../../../config.yaml');
+            $distConfigFilePath = __DIR__ . '/../../../../config.yaml';
+            $this->_distConfig = ConfigFile::createFromFile($distConfigFilePath)->toArray();
         }
-        if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
-            $this->_output->writeln('<debug>Load dist config</debug>');
-        }
+        $this->logDebug('Load dist config');
+
         $config = ArrayFunctions::mergeArrays($this->_distConfig, $initConfig);
 
         return $config;
@@ -184,11 +184,7 @@ class ConfigurationLoader
             }
 
             if ($systemWideConfigFile && file_exists($systemWideConfigFile)) {
-                if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
-                    $this->_output->writeln(
-                        '<debug>Load system config <comment>' . $systemWideConfigFile . '</comment></debug>'
-                    );
-                }
+                $this->logDebug('Load system config <comment>' . $systemWideConfigFile . '</comment>');
                 $this->_systemConfig = Yaml::parse($systemWideConfigFile);
             } else {
                 $this->_systemConfig = array();
@@ -207,7 +203,7 @@ class ConfigurationLoader
      *
      * @return array
      */
-    public function loadPluginConfig($config, $magentoRootFolder)
+    public function loadPluginConfig(array $config, $magentoRootFolder)
     {
         if ($this->_pluginConfig == null) {
             $this->_pluginConfig = array();
@@ -287,7 +283,6 @@ class ConfigurationLoader
         return str_replace(array_keys($replace), $replace, $rawConfig);
     }
 
-
     /**
      * Check if there is a user config file. ~/.n98-magerun.yaml
      *
@@ -300,17 +295,9 @@ class ConfigurationLoader
     {
         if (null === $this->_userConfig) {
             $this->_userConfig = array();
-            $homeDirectory = OperatingSystem::getHomeDir();
-            if ($homeDirectory) {
-                if (OperatingSystem::isWindows()) {
-                    $personalConfigFile = $homeDirectory . DIRECTORY_SEPARATOR . $this->_customConfigFilename;
-                } else {
-                    $personalConfigFile = $homeDirectory . DIRECTORY_SEPARATOR . '.' . $this->_customConfigFilename;
-                }
-                if (file_exists($personalConfigFile)) {
-                    $userConfig = $this->applyVariables(\file_get_contents($personalConfigFile), $magentoRootFolder);
-                    $this->_userConfig = Yaml::parse($userConfig);
-                }
+            $locator = new ConfigLocator($this->_customConfigFilename, $magentoRootFolder);
+            if ($userConfigFile = $locator->getUserConfigFile()) {
+                $this->_userConfig = $userConfigFile->toArray();
             }
         }
 
@@ -330,33 +317,23 @@ class ConfigurationLoader
      */
     public function loadProjectConfig($magentoRootFolder, $magerunStopFileFolder, array $config)
     {
-        if ($this->_projectConfig == null) {
-            $this->_projectConfig = array();
-
-            $projectConfigFile = $magentoRootFolder . DIRECTORY_SEPARATOR . 'app/etc/' . $this->_customConfigFilename;
-            if ($projectConfigFile && file_exists($projectConfigFile)) {
-                $projectConfig = $this->applyVariables(
-                    \file_get_contents($projectConfigFile),
-                    $magentoRootFolder,
-                    null
-                );
-                $this->_projectConfig = Yaml::parse($projectConfig);
-            }
-
-            $stopFileConfigFile = $magerunStopFileFolder . DIRECTORY_SEPARATOR . '.' . $this->_customConfigFilename;
-            if (!empty($magerunStopFileFolder) && file_exists($stopFileConfigFile)) {
-                $projectConfig = $this->applyVariables(
-                    \file_get_contents($stopFileConfigFile),
-                    $magentoRootFolder,
-                    null
-                );
-                $this->_projectConfig = ArrayFunctions::mergeArrays($this->_projectConfig, Yaml::parse($projectConfig));
-            }
-
-            $config = ArrayFunctions::mergeArrays($config, $this->_projectConfig);
+        if (null !== $this->_projectConfig) {
+            return ArrayFunctions::mergeArrays($config, $this->_projectConfig);
         }
 
-        return $config;
+        $this->_projectConfig = array();
+
+        $locator = new ConfigLocator($this->_customConfigFilename, $magentoRootFolder);
+
+        if ($projectConfigFile = $locator->getProjectConfigFile()) {
+            $this->_projectConfig = $projectConfigFile->toArray();
+        }
+
+        if ($stopFileConfigFile = $locator->getStopFileConfigFile($magerunStopFileFolder)) {
+            $this->_projectConfig = $stopFileConfigFile->mergeArray($this->_projectConfig);
+        }
+
+        return ArrayFunctions::mergeArrays($config, $this->_projectConfig);
     }
 
     /**
@@ -377,19 +354,16 @@ class ConfigurationLoader
             }
 
             if ($path === false) {
-                $this->_output->writeln(sprintf("<error>Plugin config file broken link '%s'</error>", $file));
+                $this->log(sprintf("<error>Plugin config file broken link '%s'</error>", $file));
                 return;
             }
         }
 
-        if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
-            $this->_output->writeln('<debug>Load plugin config <comment>' . $path . '</comment></debug>');
-        }
+        $this->logDebug('Load plugin config <comment>' . $path . '</comment>');
 
-        $localPluginConfig = \file_get_contents($path);
-        $localPluginConfig = Yaml::parse($this->applyVariables($localPluginConfig, $magentoRootFolder, $file));
-
-        $this->_pluginConfig = ArrayFunctions::mergeArrays($this->_pluginConfig, $localPluginConfig);
+        $localPluginConfigFile = ConfigFile::createFromFile($path);
+        $localPluginConfigFile->applyVariables($magentoRootFolder, $file);
+        $this->_pluginConfig = $localPluginConfigFile->mergeArray($this->_pluginConfig);
     }
 
     /**
@@ -418,5 +392,23 @@ class ConfigurationLoader
     public function getConfigurationLoaderDir()
     {
         return __DIR__;
+    }
+
+    /**
+     * @param string $message
+     */
+    private function logDebug($message)
+    {
+        if (OutputInterface::VERBOSITY_DEBUG <= $this->_output->getVerbosity()) {
+            $this->log('<debug>' . $message . '</debug>');
+        }
+    }
+
+    /**
+     * @param string $message
+     */
+    private function log($message)
+    {
+        $this->_output->writeln($message);
     }
 }